Drastically reduce CPU usage and allocations
`SerialPort.Read` takes a `count` parameter, which specifies the maximum
number of bytes to read from the serial port. If this is non-zero, then
the `Read` call will block until at least 1 byte has been received.

Previously, if there were no bytes available to read, then the code span
in a polling loop until a byte became available, burning CPU. It's more
efficient to call `Read` with a non-zero `count` parameter in this case,
which causes the thread to sleep until one or more bytes are available.

Change the `Run` method to follow the normal pattern of allocating a single
buffer, and then reading some number of bytes into it, and processing
just the bytes which were read. This also avoids the thousands of
allocations of single-byte arrays which were happening previously.

The `_sleepTime` field is now a bit redundent: I assume this was added
in an attempt to reduce CPU usage. I've left it in out of backwards
compatibility concerns, but it's not adding anything useful any more.